+++ src/cli/tools/web/server.js
@@ -1,30 +1,50 @@
 /* eslint-disable no-console */
+const axios = require('axios');
 const path = require('path');
 const fs = require('fs');
-const finalhandler = require('finalhandler');
 const http = require('http');
 const serveStatic = require('serve-static');
 const { URL } = require('url');
 
-const { ABT_NETWORKS_PATH } = require('../../../constant');
+const { ABT_NETWORKS_PATH, ABT_NETWORKS_URL } = require('../../../constant');
+
+const getRemoteNetworks = async () => {
+  const networks = [];
+  try {
+    const { data } = await axios.get(ABT_NETWORKS_URL, { timeout: 10 * 1000 });
+    if (Array.isArray(data)) {
+      networks.push(...data);
+    }
+  } catch (error) {
+    console.log(error);
+  }
+
+  return networks;
+};
 
 const responseNetworks = res => {
   try {
-    res.setHeader('content-type', 'text/javascript');
-    res.write(
-      `window.ARCBLOCK_NETWORKS=${fs
-        .readFileSync(ABT_NETWORKS_PATH)
-        .toString()
-        .trim() || JSON.stringify([])}`
-    );
-    res.end();
+    const tmp = fs
+      .readFileSync(ABT_NETWORKS_PATH)
+      .toString()
+      .trim();
+
+    const networks = tmp ? JSON.parse(tmp) : [];
+    getRemoteNetworks().then(remoteNetworks => {
+      networks.push(...remoteNetworks);
+
+      res.setHeader('content-type', 'text/javascript');
+      res.write(`window.ARCBLOCK_NETWORKS=${JSON.stringify(networks)}`);
+      res.end();
+    });
   } catch (error) {
     console.log('response networks.js failed', error);
     res.end();
   }
 };
 
-const serve = serveStatic(path.join(__dirname, '../../../../node_modules/@arcblock/forge-web'), {
+const filesPath = path.join(__dirname, '../../../../node_modules/@arcblock/forge-web');
+const serve = serveStatic(filesPath, {
   index: ['index.html', 'index.htm'],
 });
 
@@ -42,7 +62,10 @@ const server = http.createServer((req, res) => {
     console.error(error);
   }
 
-  return serve(req, res, finalhandler(req, res));
+  return serve(req, res, () => {
+    res.write(fs.readFileSync(path.join(filesPath, 'index.html')).toString());
+    res.end();
+  });
 });
 
 server.listen(process.env.FORGE_WEB_PROT);